postgres_to_es/requirements.txt
psycopg2-binary==2.9.6
pydantic==1.10.7
python-dateutil==2.8.2
six==1.16.0
typing_extensions==4.5.0
